build: pin blst fork for memory fix
